+++ src/Notifications/NotificationTemplateResponse/State.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Courier\Notifications\NotificationTemplateGetResponse;
+namespace Courier\Notifications\NotificationTemplateResponse;
 
 /**
  * The template state. Always uppercase.